requirements.txt
streamlit
easyocr
opencv-python-headless
python-pptx
pymupdf
numpy
Pillow